It's almost done. I can feel it

+++ Java/src/org/concentus/SilkDecoderControl.java
@@ -50,9 +50,9 @@ class SilkDecoderControl
     {
         Arrays.MemSet(pitchL, 0, SilkConstants.MAX_NB_SUBFR);
         Arrays.MemSet(Gains_Q16, 0, SilkConstants.MAX_NB_SUBFR);
-        Arrays.MemSet(PredCoef_Q12[0], 0, SilkConstants.MAX_LPC_ORDER);
-        Arrays.MemSet(PredCoef_Q12[1], 0, SilkConstants.MAX_LPC_ORDER);
-        Arrays.MemSet(LTPCoef_Q14, 0, SilkConstants.LTP_ORDER * SilkConstants.MAX_NB_SUBFR);
+        Arrays.MemSet(PredCoef_Q12[0], (short)0, SilkConstants.MAX_LPC_ORDER);
+        Arrays.MemSet(PredCoef_Q12[1], (short)0, SilkConstants.MAX_LPC_ORDER);
+        Arrays.MemSet(LTPCoef_Q14, (short)0, SilkConstants.LTP_ORDER * SilkConstants.MAX_NB_SUBFR);
         LTP_scale_Q14 = 0;
     }
 }
