+++ Java/src/org/concentus/SilkEncoderControl.java
@@ -74,13 +74,13 @@ class SilkEncoderControl
     void Reset()
     {
         Arrays.MemSet(Gains_Q16, 0, SilkConstants.MAX_NB_SUBFR);
-        Arrays.MemSet(PredCoef_Q12[0], 0, SilkConstants.MAX_LPC_ORDER);
-        Arrays.MemSet(PredCoef_Q12[1], 0, SilkConstants.MAX_LPC_ORDER);
-        Arrays.MemSet(LTPCoef_Q14, 0, SilkConstants.LTP_ORDER * SilkConstants.MAX_NB_SUBFR);
+        Arrays.MemSet(PredCoef_Q12[0], (short)0, SilkConstants.MAX_LPC_ORDER);
+        Arrays.MemSet(PredCoef_Q12[1], (short)0, SilkConstants.MAX_LPC_ORDER);
+        Arrays.MemSet(LTPCoef_Q14, (short)0, SilkConstants.LTP_ORDER * SilkConstants.MAX_NB_SUBFR);
         LTP_scale_Q14 = 0;
         Arrays.MemSet(pitchL, 0, SilkConstants.MAX_NB_SUBFR);
-        Arrays.MemSet(AR1_Q13, 0, SilkConstants.MAX_NB_SUBFR * SilkConstants.MAX_SHAPE_LPC_ORDER);
-        Arrays.MemSet(AR2_Q13, 0, SilkConstants.MAX_NB_SUBFR * SilkConstants.MAX_SHAPE_LPC_ORDER);
+        Arrays.MemSet(AR1_Q13, (short)0, SilkConstants.MAX_NB_SUBFR * SilkConstants.MAX_SHAPE_LPC_ORDER);
+        Arrays.MemSet(AR2_Q13, (short)0, SilkConstants.MAX_NB_SUBFR * SilkConstants.MAX_SHAPE_LPC_ORDER);
         Arrays.MemSet(LF_shp_Q14, 0, SilkConstants.MAX_NB_SUBFR);
         Arrays.MemSet(GainsPre_Q14, 0, SilkConstants.MAX_NB_SUBFR);
         Arrays.MemSet(HarmBoost_Q14, 0, SilkConstants.MAX_NB_SUBFR);
